More layers refactoring
